chore: refactor hero animations
